flask_rpi/requirements.txt
Flask
Flask-Caching
Flask-Compress
Flask-Assets
Flask-WTF
unidecode
unicode
requests
python-nmap
cups